Add Modal training; fix negative val_bpb; max_val_tokens; log_every prints
Modal:
- modal_train.py: download_data + train functions on A10G GPU
- Uses pinky-lm-cache volume for data/checkpoints
- FINEWEB_DATA_ROOT env var to redirect downloads into volume
- data/cached_challenge_fineweb.py: respect FINEWEB_DATA_ROOT override

Fixes:
- base_bytes int16 → int32 (MPS advanced indexing bug causing negative bpb)
- max_val_batches → max_val_tokens (consistent regardless of batch size)
- log_every now prints to terminal, not just wandb
- Single wandb.log() per step (was double-logging at eval steps)
- Config suffix appended to wandb run name

+++ src/tokenizer.py
@@ -45,7 +45,7 @@ class SentencePieceTokenizer:
             base_bytes_np[token_id] = len(piece.encode('utf-8'))
 
         return (
-            torch.tensor(base_bytes_np,        dtype=torch.int16, device=device),
+            torch.tensor(base_bytes_np,        dtype=torch.int32, device=device),
             torch.tensor(has_leading_space_np, dtype=torch.bool,  device=device),
             torch.tensor(is_boundary_np,       dtype=torch.bool,  device=device),
         )
